refactor: component, pages 분리
