--- front/src/components/IntroButtonBox/index.tsx
@@ -1,1 +1,0 @@
-export { default as IntroButtonBox } from './IntroButtonBox';
