--- front/src/components/Introduce/index.tsx
@@ -1,1 +1,0 @@
-export { default as Introduce } from './Introduce';
