--- front/src/components/LoginBox/index.tsx
@@ -1,1 +1,0 @@
-export { default as LoginBox } from './LoginBox';
