+++ front/src/components/SnowGlobeCanvas/Snow.tsx
@@ -1,6 +1,5 @@
-import { Vector3 } from '@react-three/fiber';
-import { useFrame } from '@react-three/fiber';
 import { useRef } from 'react';
+import { Vector3, useFrame } from '@react-three/fiber';
 import { Mesh } from 'three';
 
 interface SnowProps {
@@ -11,12 +10,14 @@ interface SnowProps {
 const Snow = (props: SnowProps) => {
   const snowRef = useRef<Mesh>(null);
   const speed = 0.05;
+
   useFrame(() => {
     if (snowRef.current) {
       if (snowRef.current.position.y <= 0) {
         snowRef.current.position.y = 22;
       }
       snowRef.current.position.y -= speed;
+
       if (
         snowRef.current.position.x ** 2 +
           (snowRef.current.position.y - 8) ** 2 +
@@ -29,6 +30,7 @@ const Snow = (props: SnowProps) => {
       }
     }
   });
+
   return (
     <mesh position={props.position} ref={snowRef}>
       <sphereGeometry args={[props.radius, 32, 16]} />
