+++ front/src/components/index.tsx
@@ -1,5 +1,2 @@
 export { SnowGlobeCanvas } from './SnowGlobeCanvas';
 export { Button } from './Button';
-export { IntroButtonBox } from './IntroButtonBox';
-export { Introduce } from './Introduce';
-export { LoginBox } from './LoginBox';
